refactored gdl package structure to new scheme
