+++ game/constructs/board/Board.groovy
@@ -1,7 +1,7 @@
 package constructs.board
 
-import gdl.GDLClause
-import gdl.HasClauses
+import gdl.clauses.GDLClause
+import gdl.clauses.HasClauses
 
 /**
  * @author Lawrence Thatcher
