+++ game/gdl/GDLConvertable.groovy
@@ -1,3 +1,5 @@
+package gdl
+
 import gdl.GDLDescription
 
 /**
