+++ game/gdl/clauses/dynamic/DynamicComponentsClause.groovy
@@ -1,7 +1,7 @@
-package gdl.dynamic
+package gdl.clauses.dynamic
 
-import gdl.AbstractClause
-import gdl.ClauseType
+import gdl.clauses.AbstractClause
+import gdl.clauses.ClauseType
 import gdl.GDLStatement
 
 /**
