+++ game/gdl/clauses/init/InitClause.groovy
@@ -1,7 +1,7 @@
-package gdl.init
+package gdl.clauses.init
 
-import gdl.AbstractClause
-import gdl.ClauseType
+import gdl.clauses.AbstractClause
+import gdl.clauses.ClauseType
 import gdl.GDLStatement
 
 /**
